Extract function into separate module
